// src/modules/customer/types/customer.types.ts
/**
 * Customer Domain Types & Constants
 * Reference: DB0 schema & Phase B4 specifications
 */

export enum CustomerErrorCode {
  CUSTOMER_NOT_FOUND = 'CUSTOMER_NOT_FOUND',
  CUSTOMER_INACTIVE = 'CUSTOMER_INACTIVE',
  CUSTOMER_SUSPENDED = 'CUSTOMER_SUSPENDED',
  ADDRESS_NOT_FOUND = 'ADDRESS_NOT_FOUND',
  ADDRESS_ACCESS_DENIED = 'ADDRESS_ACCESS_DENIED',
  INVALID_ADDRESS = 'INVALID_ADDRESS',
  INVALID_PHONE = 'INVALID_PHONE',
  INVALID_POSTAL_CODE = 'INVALID_POSTAL_CODE',
  INVALID_COORDINATES = 'INVALID_COORDINATES',
  SERVICE_NOT_FOUND = 'SERVICE_NOT_FOUND',
  FAVORITE_NOT_FOUND = 'FAVORITE_NOT_FOUND',
  REWARD_ACCOUNT_NOT_FOUND = 'REWARD_ACCOUNT_NOT_FOUND',
}

export enum CustomerAuditEventType {
  CUSTOMER_PROFILE_UPDATED = 'CUSTOMER_PROFILE_UPDATED',
  ADDRESS_CREATED = 'ADDRESS_CREATED',
  ADDRESS_UPDATED = 'ADDRESS_UPDATED',
  ADDRESS_SET_DEFAULT = 'ADDRESS_SET_DEFAULT',
  ADDRESS_DEACTIVATED = 'ADDRESS_DEACTIVATED',
  FAVORITE_ADDED = 'FAVORITE_ADDED',
  FAVORITE_REMOVED = 'FAVORITE_REMOVED',
}